feat: reorganize files based on project structure
